+++ packages/unplugin/.eslintrc.js
@@ -1,0 +1,13 @@
+module.exports = {
+  "root": true,
+  "extends": [
+    "lxsmnsyc/typescript"
+  ],
+  "parserOptions": {
+    "project": "./tsconfig.eslint.json",
+    "tsconfigRootDir": __dirname,
+  },
+  "rules": {
+    "no-param-reassign": "off"
+  }
+};
