+++ packages/unplugin/env.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="babel-plugin-forgetti" />
