+++ packages/unplugin/package.json
@@ -27,21 +27,19 @@
     "pridepack",
     "babel"
   ],
-  "name": "vite-plugin-thaler",
+  "name": "unplugin-thaler",
   "devDependencies": {
     "@types/node": "^18.13.0",
     "eslint": "^8.36.0",
     "eslint-config-lxsmnsyc": "^0.5.0",
     "pridepack": "^2.4.1",
     "tslib": "^2.5.0",
-    "typescript": "^4.9.5",
-    "vite": "^4.2.1"
+    "typescript": "^4.9.5"
   },
   "dependencies": {
-    "unplugin-thaler": "0.3.1"
-  },
-  "peerDependencies": {
-    "vite": "^3 || ^4"
+    "@rollup/pluginutils": "^5.0.2",
+    "thaler": "0.3.0",
+    "unplugin": "^1.3.1"
   },
   "scripts": {
     "prepublish": "pridepack clean && pridepack build",
@@ -57,7 +55,7 @@
     "url": "https://github.com/lxsmnsyc/thaler.git",
     "type": "git"
   },
-  "homepage": "https://github.com/lxsmnsyc/thaler/tree/main/packages/vite",
+  "homepage": "https://github.com/lxsmnsyc/thaler/tree/main/packages/unplugin",
   "bugs": {
     "url": "https://github.com/lxsmnsyc/thaler/issues"
   },
